refactor: Rename module command to rustup_command
